requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
boto3==1.34.0
Werkzeug==2.3.7